Move cli tests to their own subdir
