+++ tests/cli/mod.rs
@@ -1,0 +1,2 @@
+mod build;
+mod test;
