+++ tests/cli_tests.rs
@@ -1,0 +1,2 @@
+mod cli;
+mod dummy_book;
